cleanup slides fixtures
